Change strings to format strings; Add exception handling
Changed strings to format strings. i.e. %s for formatting string variables into text, %d for ints, and %f for floats. %0.2f passes floats in rounded/truncated to 2 decimal places. 

Added exception handling for value errors when entering not int values, it's better to ask for forgiveness than to ask for permission. Alternatively, the float() method may be applied to a string rather than int() in order to allow for non-whole values to be entered.